+++ packages/patient-portal/src/views/LoginView.tsx
@@ -5,6 +5,20 @@ import { useLocation } from 'react-router';
 import { useLogin } from '@api/mutations';
 import { TextField } from '../components/TextField';
 
+const Card = styled(Paper)`
+  margin: 100px auto;
+  display: block;
+  padding: 22px;
+  min-width: 300px;
+  width: 520px;
+  max-width: 100%;
+  text-align: center;
+  border-radius: 10px;
+  box-shadow: none;
+  border: none;
+`;
+
+
 export const LoginView = () => {
   const { mutate: login } = useLogin();
   const location = useLocation();
